fast api/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
torch
transformers
numpy
jinja2
python-jose[cryptography]
